refactor: domains에서 page분리
