+++ src/pages/CreatePost/CreateMainPage/CreateMain.style.ts
