+++ src/pages/CreatePost/SearchPage/Search.style.ts
