+++ src/pages/CreatePost/SetDatePage/index.ts
@@ -1,0 +1,1 @@
+export { SetDatePage } from './SetDatePage';
