+++ src/pages/CreatePost/SetPlaceMapPage/SetPlaceMap.style.ts
