+++ src/pages/CreatePost/SetPlaceMapPage/index.ts
@@ -1,0 +1,1 @@
+export { SetPlaceMapPage } from './SetPlaceMapPage';
