+++ src/pages/CreatePost/SetPlacePage/SetPlace.style.ts
