+++ src/pages/PostDetail/page.style.ts
